# Conflicts:
#	backend-rest-api/src/main/java/com/nitsoft/ecommerce/service/categories/CategoriesServiceImpl.java

+++ backend-rest-api/src/main/java/com/nitsoft/ecommerce/repository/CompanyRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.repository.query.Param;
 public interface CompanyRepository extends CrudRepository<Company, Integer> {
 
     @Query("SELECT c FROM Company c")
-    Iterable<Company> findAll(Integer companyId);
+    Iterable<Company> findAll(long companyId);
 
     @Query("SELECT c FROM Company c WHERE c.companyId = :companyId")
-    Company findByCompanyId(@Param("companyId") Integer companyId);
+    Company findByCompanyId(@Param("companyId") long companyId);
     
 }
